requirements.txt
astor==0.8.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dlib==19.21.1
dominate==2.6.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
fvcore==0.1.5.post20210727
h5py==2.10.0
html5lib==0.9999999
hyperopt==0.1.2
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
iopath==0.1.9
joblib==1.0.1
json-tricks==3.15.5
Keras==2.1.2
kiwisolver==1.3.1
lpips==0.1.3
Markdown==3.3.4
matplotlib==3.3.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
netifaces==0.11.0
networkx==2.5
nni==2.2
numpy==1.19.2
olefile==0.46
opencv-python==4.5.1.48
Pillow==8.1.0
pip==21.0.1
portalocker==2.3.0
prettytable==2.1.0
progress==1.5
protobuf==3.14.0
psutil==5.8.0
pycparser==2.20
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
PythonWebHDFS==0.2.3
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
responses==0.13.3
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
schema==0.7.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.2
setuptools==52.0.0.post20210125
simplejson==3.17.2
six==1.15.0
tabulate==0.8.9
tensorboardX==2.1
tensorflow==1.3.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0
torch-fidelity==0.2.0
torchsummary==1.5.1
torchvision==0.2.1
tqdm==4.56.0
urllib3==1.26.5
wcwidth==0.2.5
websockets==9.1
Werkzeug==1.0.1
wheel==0.36.2
yacs==0.1.8
zipp==3.4.0
